server/requirements.txt
google-search-results
fastapi
fastapi[standard]
typing
requests
python-dotenv
uvicorn
pydantic
google-genai
sqlmodel
sqlalchemy